+++ libindi/drivers/ccd/ccd_simulator.cpp
@@ -987,21 +987,21 @@ int CCDSim::AddToPixel(INDI::CCDChip *targetChip, int x, int y, int val)
     return drew;
 }
 
-IPState CCDSim::GuideNorth(float v)
+IPState CCDSim::GuideNorth(uint32_t v)
 {
-    guideNSOffset    += v / 1000 * GuideRate / 3600;
+    guideNSOffset    += v / 1000.0 * GuideRate / 3600;
     return IPS_OK;
 }
 
-IPState CCDSim::GuideSouth(float v)
+IPState CCDSim::GuideSouth(uint32_t v)
 {
-    guideNSOffset    += v / -1000 * GuideRate / 3600;
+    guideNSOffset    += v / -1000.0 * GuideRate / 3600;
     return IPS_OK;
 }
 
-IPState CCDSim::GuideEast(float v)
+IPState CCDSim::GuideEast(uint32_t v)
 {
-    float c   = v / 1000 * GuideRate;
+    float c   = v / 1000.0 * GuideRate;
     c   = c/ 3600.0 / 15.0;
     c   = c/ (cos(currentDE * 0.0174532925));
 
@@ -1010,9 +1010,9 @@ IPState CCDSim::GuideEast(float v)
     return IPS_OK;
 }
 
-IPState CCDSim::GuideWest(float v)
+IPState CCDSim::GuideWest(uint32_t v)
 {
-    float c   = v / -1000 * GuideRate;
+    float c   = v / -1000.0 * GuideRate;
     c   = c/ 3600.0 / 15.0;
     c   = c/ (cos(currentDE * 0.0174532925));
