+++ libindi/drivers/telescope/ieqpro.h
@@ -82,10 +82,10 @@ class IEQPro : public INDI::Telescope, public INDI::GuiderInterface
     void mountSim();
 
     // Guide
-    virtual IPState GuideNorth(float ms) override;
-    virtual IPState GuideSouth(float ms) override;
-    virtual IPState GuideEast(float ms) override;
-    virtual IPState GuideWest(float ms) override;
+    virtual IPState GuideNorth(uint32_t ms) override;
+    virtual IPState GuideSouth(uint32_t ms) override;
+    virtual IPState GuideEast(uint32_t ms) override;
+    virtual IPState GuideWest(uint32_t ms) override;
 
   private:
     /**
